CLSTM_Depth_Estimation-master/requirements.txt
matplotlib==3.0.3
torchvision==0.2.2
torch==1.0.0
numpy==1.16.2
opencv_python==4.1.0.25
Pillow==6.2.0
